util/utiltrencodings duplicates moved and removed
