+++ src/netbase.cpp
@@ -13,6 +13,7 @@
 
 #include "netbase.h"
 #include "util.h"
+#include "utilstrencodings.h"
 #include "sync.h"
 #include "hash.h"
