+++ src/script/script.h
@@ -18,6 +18,7 @@
 #include "pubkey.h"
 #include "bignum.h"
 #include "util.h"
+#include "utilstrencodings.h"
 #include "anon/stealth/stealth.h"
 
 typedef std::vector<unsigned char> valtype;
