+++ src/sync.cpp
@@ -3,9 +3,13 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <boost/foreach.hpp>
+#include <boost/thread.hpp>
+
+#include <stdio.h>
 
 #include "sync.h"
 #include "util.h"
+#include "utilstrencodings.h"
 
 #ifdef DEBUG_LOCKCONTENTION
 void PrintLockContention(const char* pszName, const char* pszFile, int nLine)
