+++ src/timedata.cpp
@@ -9,6 +9,7 @@
 #include "sync.h"
 #include "ui_interface.h"
 #include "util.h"
+#include "utilstrencodings.h"
 
 using namespace std;
