+++ src/wallet/db.cpp
@@ -15,6 +15,7 @@
 #include "addrman.h"
 #include "hash.h"
 #include "util.h"
+#include "utilstrencodings.h"
 
 using namespace std;
 using namespace boost;
