+++ src/wallet/wallet.h
@@ -20,6 +20,7 @@
 #include "script/script.h"
 #include "ui_interface.h"
 #include "util.h"
+#include "utilstrencodings.h"
 #include "anon/stealth/stealth.h"
 
 const CAmount MIN_TX_FEE = 10000; // 0.00001 DRKSLK Minimum Transaction Fee
